src/requirements.txt
alabaster==0.7.12
arrow==1.2.3
astroid==2.11.7
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==4.1.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.6
comm==0.1.2
cookiecutter==1.7.3
cryptography==38.0.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docutils==0.18.1
entrypoints==0.4
fastjsonschema==2.16.2
flake8==4.0.1
Flask==2.2.2
flit_core==3.6.0
idna==3.4
imagesize==1.4.1
importlib-metadata==4.11.3
inflection==0.5.1
install==1.3.5
intervaltree==3.1.0
ipykernel==6.19.2
ipython==7.31.1
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.16.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab-pygments==0.1.2
keyring==23.4.0
lazy-object-proxy==1.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.24.3
numpydoc==1.5.0
packaging==22.0
pandas==2.0.1
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.0.1
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.14.5
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
python-slugify==5.0.2
pytz==2022.7
pywin32==305.1
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
rope==0.22.0
Rtree==1.0.1
setuptools==67.7.2
sip==6.6.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.3
spyder-kernels==2.3.3
text-unidecode==1.3
textdistance==4.2.1
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2023.3
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.14
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whatthepatch==1.0.2
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.14.1
yapf==0.31.0
zipp==3.11.0
